backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
duckdb>=1.0.0
pydantic==2.11.10
python-multipart==0.0.20
python-dateutil==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
boto3>=1.34.0
sentry-sdk[fastapi]>=2.0.0
sendgrid>=6.11.0
python-dotenv>=1.0.0